+++ app/Livewire/Planning/SearchString/SearchTerm.php
@@ -11,9 +11,13 @@ use App\Models\Synonym as SynonymModel;
 use App\Utils\ActivityLogHelper as Log;
 use App\Utils\ToastHelper;
 use Illuminate\Support\Facades\Http;
+use App\Traits\ProjectPermissions;
 
 class SearchTerm extends Component
 {
+
+    use ProjectPermissions;
+
     private $translationPath = 'project/planning.search-string.term.livewire';
     private $toastMessages = 'project/planning.search-string.term.livewire.toasts';
 
@@ -114,6 +118,11 @@ class SearchTerm extends Component
      */
     public function submit()
     {
+
+        if (!$this->checkEditPermission($this->toastMessages . '.denied')) {
+            return;
+        }
+
         $this->validate();
 
         $updateIf = [
@@ -156,6 +165,11 @@ class SearchTerm extends Component
      */
     public function edit(string $termId)
     {
+
+        if (!$this->checkEditPermission($this->toastMessages . '.denied')) {
+            return;
+        }
+
         $this->currentTerm = SearchTermModel::findOrFail($termId);
         $this->description = $this->currentTerm->description;
         $this->form['isEditing'] = true;
@@ -171,6 +185,11 @@ class SearchTerm extends Component
      */
     public function delete(string $termId)
     {
+
+        if (!$this->checkEditPermission($this->toastMessages . '.denied')) {
+            return;
+        }
+
         try {
             $currentTerm = SearchTermModel::findOrFail($termId);
             $currentTerm->delete();
@@ -198,20 +217,13 @@ class SearchTerm extends Component
 
     public function addSynonyms()
     {
-        if (empty($this->termId['value'])) {
-            $this->toast(
-                message: __('O termo é obrigatório.'),
-                type: 'error'
-            );
+
+        if (!$this->checkEditPermission($this->toastMessages . '.denied')) {
             return;
         }
 
-        if (empty($this->synonym)) {
-            $this->toast(
-                message: __('O sinônimo não pode estar vazio.'),
-                type: 'error'
-            );
-            return;
+        if (!$this->termId['value'] || !$this->synonym) {
+            $this->addError('termId', 'The term id is required');
         }
 
         $updateIf = [
@@ -254,6 +266,11 @@ class SearchTerm extends Component
      */
     public function editSynonym($termId)
     {
+
+        if (!$this->checkEditPermission($this->toastMessages . '.denied')) {
+            return;
+        }
+
         $this->currentSynonym = SynonymModel::findOrFail($termId);
         $this->currentTerm = SearchTermModel::findOrFail($this->currentSynonym->id_term)->first();
         $this->synonym = $this->currentSynonym->description;
@@ -266,6 +283,11 @@ class SearchTerm extends Component
      */
     public function deleteSynonym(string $termId)
     {
+
+        if (!$this->checkEditPermission($this->toastMessages . '.denied')) {
+            return;
+        }
+        
         try {
             $currentSynonym = SynonymModel::findOrFail($termId);
             $currentSynonym->delete();
